fix compilation under ARM device.
qReal translate to different types on different architectures.
Consensus is that the current state of affairs will be kept for qt 5.0
too [1].

Luckily the function affected "SearchFieldRangeInt" is only used to
store exposure times [2] where 6 precision digit suffice.

So we add functions to write <float> and cast to <qReal> where needed
and be done with that.

[1] https://bugreports.qt-project.org/browse/QTBUG-23758
[2] ChangeLog 2008-07-07 For use with the exposuretime field ...
BUG: 296141

P.S. CustomStepsDoubleSpinBox could be converted to
     CustomStepsFloatSpinBox but seem not worth the effort.
     Marcel hope you're ok with this

CCMAIL: marcel.wiesweg@gmx.de

+++ utilities/searchwindow/searchfields.cpp
@@ -1209,7 +1209,7 @@ void SearchFieldRangeInt::write(SearchXmlWriter& writer)
 
             if (m_reciprocal)
             {
-                writer.writeValue(QList<double>() << m_secondBox->fractionMagicValue() << m_firstBox->fractionMagicValue());
+                writer.writeValue(QList<float>() << m_secondBox->fractionMagicValue() << m_firstBox->fractionMagicValue());
             }
             else
             {
